Rename package
Rename package from capabilitydiscoverymode to mode, to be more concise.

#23
