+++ springcloud/src/main/java/org/axonframework/extensions/springcloud/commandhandling/mode/IgnoreListingCapabilityDiscoveryMode.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.axonframework.extensions.springcloud.commandhandling.capabilitydiscoverymode;
+package org.axonframework.extensions.springcloud.commandhandling.mode;
 
 import org.axonframework.commandhandling.distributed.CommandMessageFilter;
 import org.slf4j.Logger;
